+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CLI: Add entity instances to flow types
+- Server CLI: Print GraphQL schemas
 - GraphQL: Add entity instances to flow types
 - GraphQL: Sort fields and enums
 - Dynamic Graph: Sort components, entity types, relation types and properties